整合es和使用redis做文章查询缓存

+++ paicoding-service/src/main/java/com/github/paicoding/forum/service/article/repository/dao/ArticleDao.java
@@ -466,7 +466,7 @@ public class ArticleDao extends ServiceImpl<ArticleMapper, ArticleDO> {
                 .count();
     }
 
-    public List<ArticleDO> selectByIds(List<Integer> ids) {
+    public List<ArticleDO> selectByIds(List<Long> ids) {
 
         List<ArticleDO> articleDOS = baseMapper.selectBatchIds(ids);
         return articleDOS;
